fix: move plot_imputation to eda
